+++ base_tier_validation/__manifest__.py
@@ -15,6 +15,7 @@
     "depends": ["mail"],
     "data": [
         "data/mail_data.xml",
+        "data/cron_data.xml",
         "security/ir.model.access.csv",
         "views/tier_definition_view.xml",
         "views/tier_review_view.xml",